requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
